Always set IsProductComponent to true
Since this VSIX ships in VS, we should always set this to true.  This
allows a private build of the VSIX to install over the one included in
VS.

The tradeoff is that a privately built VSIX will not be able to install
on a VS instance which did not install LibraryManager already.