Enable roll-forward on major versions for CLI tool
This enables the package to continue to work on images which only have versions newer than netcoreapp2.1.

In the future if a breaking change is made that breaks us on roll-forward, we'll have to multi-target the build to that framework.  We shouldn't have to do this for most releases though.