Make build task messages not important by default
Currently, we show all libman build messages as Important, which means
that every detail comes through even when using /verbosity:minimal.

With this change, most messages will be shown only in normal/detailed
verbosity.  We will still show the opening and closing lines for when we
start and complete restore for minimal verbosity.