Explicity include runtime dependencies in Build package (#575)
For netstandard project, the default behavior does not copylocal
dependencies from reference packages, and does not include them in the
package.  For Newtonsoft.Json, the dependency was resolved at runtime
from the SDK install path, so we never noticed it was missing.

Since we added Minimatch.dll for the file globs, it is not being
resolved when building through the .NET sdk (dotnet build).  This is
causing the build to fail with a LIB000 error (unknown failure).

The fix is to be more precise about what files we do include in the
package.  This also addresses the inconsistency about having these
assemblies in the build output path when signing package contents.