Update dependencies to the latest, convert NuGet to CPVM
This change converts to using the Central Package Version Management
feature in NuGet, enforcing that all PackageReference versions are
tracked in one central place.

I also bumped up the dependencies for most of the packages that we rely
upon.

Code changes are solely to address new analyzer issues from updating to
the latest FxCopAnalyzers package.