Ensure manifest restore uses matching dependencies
The IDependencies includes information about the working directory for
restore.  We were incorrectly using a saved instance of this, but would
reset it whenever a new TextView was opened for a libman.json file.
This would set the cached working directory to the last opened
libman.json file, and changes to any other libman.json file would then
restore in that folder.

Addresses #638