Align combobox popup with left edge
It's been a long-time issue that the completion popup on the search box
is not aligned with the left side of the completion window.  Originally
it was attempting to align with the cursor position, but the offsets
were calculated very incorrectly.

Aligning with the caret is hard, and also not entirely desirable.
Ideally, we would align with an inflection point for the completion
session (e.g. on for library ID, it would either align with the name or
with the start of the version; for target location, it would align with
the current path segment).

This is hard to do generically since we have no idea where the
inflection points are.  So, settling for left-alignment is still better
than being oddly not quite right aligned.

This also removes the attempt to move the offset with cursor movement,
as it was exacerbating the odd alignment.

Addresses #226