Remove "..." for past tense strings
The ellipsis should be used to indicate things which are still in
progress, not things which have completed.  In most cases, this was used
in a string like "X started...".

It would be acceptable to use "Starting X..." as that is the present
progressive tense, but we can also simply remove the ellipsis without
having to re-translate a bunch of strings.

Addresses #212