Ensure that libraries added via wizard set the provider (#657)
The default behavior when adding a library, if the defaultProvider was
not yet set, was to promote the new library's provider as the default.
This requires multiple edits to the file.  When the document is open in
the editor, we only calculate the edit for inserting the library.

This change does imply that adding via the UI wizard will have a
different output than using the CLI.  At least we'll no longer generate
an invalid manifest.  We may want to consider a more advanced edit
generation such that we apply both edits instead.

Addresses #393.