Update unit tests to target net6.0 (#675)
As netcore2.1 is out of support, these tests fail to run on the CI
agents.  Updating to net6.0 will allow the tests to continue running.

Updating the src projects to net6.0 would also mean that they would stop
functioning for customers who did not have net6.0 installed.  This can
be a concern for both the MSBuild package and the CLI.  Such a change
should only be made when we are bumping the major version.