Patch-up test expecatations to match the real world
This simply makes the tests pass by recognizing the current latest version of JQuery, and not validating the default file (as it may have changed since the version being used in the test - in this case, knockout).

There was a lot of test refactoring to make the tests independent from live (i.e. ever changing) API feeds that happened after the VS2017 snap.  For now, this is the quickest path to a working build.