Generalize project system interactions to CPS
We can share behavior across all CPS projects for handling files additions/removals.  CPS is much more efficient for bulk edits - they work from the file system automatically, so we don't have to do batched edits on the UI thread in VS. This is something we should take advantage of for as many project types as possible, not limited to DotNetCoreWeb.

Addresses #710