Commit libman JSON schema
This has lived solely in SchemaStore, which being its own separate project, has led to unintended and unmonitored changes to the libman schema.  It even led to a significant break in the schema at one point.

Committing here to be the source of truth, since the libman schema should be owned by this project.  This also allows future schema changes to be committed and reviewed in tandem with future changes to libman, instead of being across different repos.