Introduce installation goal state
The existing LibraryInstallationState abstraction is not flexible about mapping files to their source on an individual basis.  It presumes that all files installed by a library retain the same folder hierarchy as the library they originate from.

In order to allow more flexibility, this change adds a new abstraction to wrap all installed files to their individual source files.  This eliminates the prior assumption, allowing for flexibility to install files to a different folder hierarchy than their originating library structure.

This flexibility also allows an important new concept: each destination file must be unique (to avoid collisions), but the sources are not so constrained.  A single source file may be installed to multiple destinations; by creating a mapping of destination-to-source, we can easily allow this.

This change prepares for a new feature to allow more granular mappings of files within a library.