Copy sources for Minimatch library directly
This library provides the glob expansion logic used in LibMan.  However,
the package published to nuget.org is based on a .NET Standard
prerelease version, which causes logs of cascading package version
conflicts sometimes.  We've suppressed those in some cases, but it still
can arise.

This change takes a direct copy of the source code for this type.
Copyright attributions is placed in the file, as is a link back to the
source. It hasn't changed in 11 years, it's unlikely to change again
(although now we could if we need to).  The original was licensed under
the MIT license.