Schema changes to support fileMappings
I updated the schema to draft-07 (the newest VS supports) in order to use the if-else logic.  This allows VS to detect that fileMappings are used and then apply a validation error that `"version"` must be set to 3.0.  (Trying to do it with an anyOf like for destination/defaultDestination would give a fairly ambiguous message instead.)