Fix build targets that were incorrectly updated
During the upgrade to target .NET 8, any references to prior targeted
frameworks were removed.  In this case, the hardcoded framework path was
replaced with $(NetFxTFM) which is defined in the libman build, but not
in the consumer's build (where this .targets file is evaluated).  It
should not have been replaced here.

I've reset it back to hard-coded values, and updated the frameworks
currently targeted by the project (net481 and net8.0).