Fix RestoreTask tallying of written files
During the switch to using LibraryInstallationGoalState, we changed the behavior of taking a LibraryInstallationState and "updating" it to reflect the published files (this transform would expand the Files from globs to a full list).  That's inherently done by GoalState now, which also takes into account the FileMappings, and produces a set of the destination files.

This was somehow overlooked in this build task, which caused it to throw a NRE when it was looking at the LibraryInstallationState (now that it wasn't being "updated", the Files were empty in many cases).

This fix re-computes each successful LibraryInstallationState into the LibraryInstallationGoalState, which contains the list of files that were installed by that library.