Fix FileSystemProvider validation during rename
Because FileSystemProvider allows specifying a file as the library and a different file name as the destination (i.e. rename scenario), it was failing during goal state validation.  However, in the FSP case, this is always valid - when a single file is specified, it can be mapped to one output file of any name.