Address cross-project contamination when restoring in VS
During the restore-on-save operation, we were using a cached instance of the "before" manifest, which was used to clean up file deltas.  A side effect of this was that we used the working folder for that cached manifest.

If the user makes changes to a second libman.json, then it cleans the delta between that file and the previous libman.json in its folder.  That typically would appear as deleting the files out of one project when making changes the libman.json file in a different project.

The fix is straightforward: cache the instance in a per-document form instead of a single stored value.